🐛 Fix: 디자인 및 오류 수정
