+++ src/app/main/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from 'next';
 import '@/style/globals.css';
+import Header from '@/components/domains/landing/header/Header';
 
 export const metadata: Metadata = {
   title: '메인 페이지',
@@ -11,5 +12,10 @@ export default function MainPageLayout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  return <div>{children}</div>;
+  return (
+    <>
+      <Header />
+      <section className="relative flex w-full h-full max-w-[1440px] mx-auto my-10">{children}</section>
+    </>
+  );
 }
