+++ src/components/commons/sidebar/Sidebar.tsx
@@ -12,11 +12,11 @@ const MOCK_USER = {
 };
 
 const Sidebar = () => {
-  const [isOpen, setIsOpen] = useState(true);
+  const [isOpen, setIsOpen] = useState(false);
 
   return (
     <div
-      className={`m-10 bg-white rounded-lg shadow ${isOpen ? 'w-[200px]' : 'w-[60px]'} duration-500 text-gray-900 px-4`}>
+      className={`h-dvh bg-white rounded-[8px] border border-gray-89 ${isOpen ? 'w-[200px]' : 'w-[60px]'} duration-500 text-gray-900 px-4`}>
       <div className={`py-3 flex mt-5 ${isOpen ? 'justify-end' : 'justify-center'}`}>
         <HiMenuAlt3 size={25} className="cursor-pointer" onClick={() => setIsOpen(!isOpen)} />
       </div>
